requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.1
torch==1.8.1
torchvision==0.9.1
streamlit
